Added ORM persistence listener
