+++ composer.json
@@ -25,7 +25,8 @@
     "suggest": {
         "guzzlehttp/guzzle": "If you want to integrate with slack you could use guzzle to send your data to slack",
         "excelwebzone/recaptcha-bundle": "Install this bundle if you want a recaptcha in your contact form",
-        "symfony/validator": "Install the symfony validation component if you want validation of your contact model class"
+        "symfony/validator": "Install the symfony validation component if you want validation of your contact model class",
+        "doctrine/orm": "Install Doctrine ORM to persist your contact form in a relational database"
     },
     "autoload": {
         "psr-4": {
