+++ src/Model/Contact.php
@@ -17,115 +17,7 @@ namespace Kristofvc\Contact\Model;
  *
  * @author Kristof Van Cauwenbergh <kristof.vancauwenbergh@gmail.com>
  */
-final class Contact implements ContactInterface
+final class Contact extends AbstractContact
 {
-    /**
-     * @var string
-     */
-    private $name;
 
-    /**
-     * @var string
-     */
-    private $email;
-
-    /**
-     * @var string
-     */
-    private $recaptcha;
-
-    /**
-     * @var string
-     */
-    private $message;
-
-    /**
-     * @var \DateTime
-     */
-    private $createdAt;
-
-    public function __construct()
-    {
-        $this->createdAt = new \DateTime();
-    }
-
-    /**
-     * @return string
-     */
-    public function getEmail()
-    {
-        return $this->email;
-    }
-
-    /**
-     * @param string $email
-     */
-    public function setEmail($email)
-    {
-        $this->email = $email;
-    }
-
-    /**
-     * @return string
-     */
-    public function getMessage()
-    {
-        return $this->message;
-    }
-
-    /**
-     * @param string $message
-     */
-    public function setMessage($message)
-    {
-        $this->message = $message;
-    }
-
-    /**
-     * @return string
-     */
-    public function getName()
-    {
-        return $this->name;
-    }
-
-    /**
-     * @param string $name
-     */
-    public function setName($name)
-    {
-        $this->name = $name;
-    }
-
-    /**
-     * @return string
-     */
-    public function getRecaptcha()
-    {
-        return $this->recaptcha;
-    }
-
-    /**
-     * @param string $recaptcha
-     */
-    public function setRecaptcha($recaptcha)
-    {
-        $this->recaptcha = $recaptcha;
-    }
-
-    /**
-     * @return string
-     */
-    public function getSubject()
-    {
-        return 'Contact by ' . $this->getName();
-    }
-
-    /**
-     * @return \DateTime
-     */
-    public function getCreatedAt()
-    {
-        return $this->createdAt;
-    }
 }
