+++ src/Resources/config/doctrine/AbstractContact.orm.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
+                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
+        http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+
+    <mapped-superclass name="Kristofvc\Contact\Model\AbstractContact">
+
+        <field name="name" type="string" length="100" />
+        <field name="email" type="string" length="255" />
+        <field name="message" type="text" />
+        <field name="createdAt" column="created_at" type="datetime" nullable="true" />
+
+    </mapped-superclass>
+</doctrine-mapping>
